// frontend/src/types/member.d.ts
export interface FamilyMember {
    id?: string;
    name: string;
    relation: string;
    age?: number;
    gender?: string;
    isAlive?: boolean;
    hasDiabetes?: boolean;
    diabetesDetails?: string;
    hasHeartDisease?: boolean;
    heartDiseaseDetails?: string;
    hasCancer?: boolean;
    cancerDetails?: string;
    otherConditions?: string;
    emailId?: string;
    parentId?: string;
  }
  